@@ -26,26 +26,29 @@
     "d3": "3.5.6",
     "flux": "2.1.1",
     "globalize": "1.x",
-    "griddle-react": "0.4.x",
-    "history": "1.17.x",
+    "griddle-react": "0.6.x",
+    "history": "2.1.2",
     "moment": "2.10.6",
     "object-assign": "3.0.0",
-    "react": "0.14.x",
-    "react-dom": "0.14.x",
+    "react": "15.x",
+    "react-dom": "15.x",
     "react-notification-system": "0.2.x",
-    "react-router": "1.0.x",
-    "react-skylight": "0.2.0",
-    "react-widgets": "3.0.x",
+    "react-router": "2.8.x",
+    "react-skylight": "0.4.1",
+    "react-widgets": "3.4.x",
     "react-widgets-globalize": "latest",
     "underscore": "1.8.3"
   },
   "devDependencies": {
-    "babelify": "6.3.0",
-    "browserify": "11.0.1",
+    "babel-core": "6.x",
+    "babelify": "8.x",
+    "babel-preset-es2015": "6.x",
+    "babel-preset-react": "6.x",
+    "browserify": "16.x",
     "concurrently": "2.0.0",
     "jest-cli": "0.5.4",
-    "reactify": "1.1.1",
-    "uglifyify": "3.0.1",
-    "watchify": "3.4.0"
+    "reactify": "1.x",
+    "uglifyify": "4.0.x",
+    "watchify": "3.x"
   }
 }
